+++ tests/data/google/raises/cases.py
@@ -226,3 +226,39 @@ class B:
             raise a.b.c.d.e.f.g.WhoseException.h.i.j.k
         else:
             pass
+
+    def func17(self) -> None:
+        """
+        It should pass.
+
+        Raises:
+            AssertionError: every time, without a message.
+        """
+        assert False
+
+    def func18(self) -> None:
+        """
+        It should pass.
+
+        Raises:
+            AssertionError: every time, with a message.
+        """
+        assert False, 'False'
+
+    def func19(self) -> None:
+        """
+        Should fail, expects `AssertionError`.
+
+        Returns:
+            None
+        """
+        assert False
+
+    def func20(self) -> None:
+        """
+        Should fail, expects `AssertionError`.
+
+        Raises:
+            AssertionError123: every time, with a message.
+        """
+        assert False