+++ tests/test_main.py
@@ -805,7 +805,7 @@ def testRaises(style: str, skipRaisesCheck: bool) -> None:
         style=style,
     )
     expected0 = [
-        'DOC501: Method `B.func1` has "raise" statements, but the docstring does not '
+        'DOC501: Method `B.func1` has raise/assert statements, but the docstring does not '
         'have a "Raises" section',
         'DOC503: Method `B.func1` exceptions in the "Raises" section in the docstring '
         'do not match those in the function body. Raised exceptions in the docstring: []. '
@@ -819,7 +819,7 @@ def testRaises(style: str, skipRaisesCheck: bool) -> None:
         'are not "raise" statements in the body',
         'DOC502: Method `B.func9a` has a "Raises" section in the docstring, but there '
         'are not "raise" statements in the body',
-        'DOC501: Function `inner9a` has "raise" statements, but the docstring does '
+        'DOC501: Function `inner9a` has raise/assert statements, but the docstring does '
         'not have a "Raises" section',
         'DOC503: Function `inner9a` exceptions in the "Raises" section in the '
         'docstring do not match those in the function body. Raised exceptions in the '
@@ -840,7 +840,18 @@ def testRaises(style: str, skipRaisesCheck: bool) -> None:
         'docstring do not match those in the function body. Raised exceptions in the '
         "docstring: ['CustomError']. Raised exceptions in the body: "
         "['exceptions.m.CustomError'].",
+        'DOC501: Method `B.func19` has raise/assert statements, but the docstring does not '
+        'have a "Raises" section',
+        'DOC503: Method `B.func19` exceptions in the "Raises" section in the '
+        'docstring do not match those in the function body. Raised exceptions in the '
+        "docstring: []. Raised exceptions in the body: ['AssertionError (implicitly "
+        "from the `assert` statement)'].",
+        'DOC503: Method `B.func20` exceptions in the "Raises" section in the '
+        'docstring do not match those in the function body. Raised exceptions in the '
+        "docstring: ['AssertionError123']. Raised exceptions in the body: "
+        "['AssertionError (implicitly from the `assert` statement)'].",
     ]
+
     expected1 = []
     expected = expected1 if skipRaisesCheck else expected0
     assert list(map(str, violations)) == expected
@@ -867,7 +878,7 @@ def testRaisesPy310plus(style: str, skipRaisesCheck: bool) -> None:
         style=style,
     )
     expected0 = [
-        'DOC501: Method `B.func10` has "raise" statements, but the docstring does not '
+        'DOC501: Method `B.func10` has raise/assert statements, but the docstring does not '
         'have a "Raises" section',
         'DOC503: Method `B.func10` exceptions in the "Raises" section in the '
         'docstring do not match those in the function body. Raised exceptions in the '